% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{fitcount}
\alias{fitcount}
\title{Ajuste de una distribucion de frecuencias de datos de un conteo}
\usage{
fitcount(X)
}
\arguments{
\item{X}{Vector numerico de valores enteros correspondiente a datos muestrales de un conteo}
}
\value{
Retorna los parametros de la distribucion Binomial Negativa y Poisson luego de aplicar el metodo de Maxima Verosimilitud a los datos muestrales de conteo. Ademas, presenta los siguientes graficos: la distribucion de frecuencias de los datos analizados, la distribucion de frecuencias vs distribucion Binomial Negativa, la distribucion de frecuencias vs distribucion de Poisson
}
\description{
Ajuste de una distribucion de frecuencias de datos de un conteo
}
\examples{
View(data1)
Par<-fitcount(data1$Ascaris.worm.burden)
}
