% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.R
\name{freq}
\alias{freq}
\title{Distribucion de frecuencias de datos de un conteo}
\usage{
freq(X)
}
\arguments{
\item{X}{Vector numerico de valores enteros correspondiente a datos muestrales de un conteo}
}
\value{
Retorna un vector de la distribucion de frecuencia de los datos muestrales analizados y un grafico de esta distribucion
}
\description{
Distribucion de frecuencias de datos de un conteo
}
\examples{
View(data1)
Fx<-freq(data1$Ascaris.worm.burden)
}
